# Improve performance of Complex#{+,-,*,/,**,abs2}