@nurse 2017-05-19
I don't know complex's performance improvement is worth increasing code complexity.